PAYPAY/requirements.txt
ansicon==1.89.0
anyio==4.9.0
ariadne==0.26.2
ariadne-django==0.3.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
awsebcli==3.24
bcrypt==4.3.0
blessed==1.21.0
botocore==1.38.17
cement==2.10.14
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
clr_loader==0.2.7.post0
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.2
daphne==4.2.0
decorator==5.2.1
Deprecated==1.2.18
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
fabric==3.2.2
graphql-core==3.2.5
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
invoke==2.2.0
jinxed==1.3.0
jmespath==1.0.1
packaging==24.2
paramiko==3.5.1
pathspec==0.12.1
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pypiwin32==223
pytest==8.3.5
pytest-asyncio==1.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pywin32==310
PyYAML==6.0.2
requests==2.32.3
semantic-version==2.10.0
service-identity==24.2.0
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
termcolor==2.5.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
wrapt==1.17.2
zope.interface==7.2
